+++ tests/testthat/test_wdpa_fetch.R
@@ -4,7 +4,7 @@ test_that("country name", {
   skip_on_cran()
   skip_if_not(curl::has_internet())
   skip_on_github_workflow("Windows")
-  x <- suppressWarnings(wdpa_fetch("Liechtenstein", force = TRUE, wait = TRUE))
+  x <- suppressWarnings(wdpa_fetch("Liechtenstein", wait = TRUE))
   expect_is(x, "sf")
   expect_true(all(x$ISO3 == "LIE"))
 })
@@ -13,8 +13,7 @@ test_that("ISO3", {
   skip_on_cran()
   skip_if_not(curl::has_internet())
   skip_on_github_workflow("Windows")
-  x <- suppressWarnings(wdpa_fetch(
-    "LIE", force = TRUE, wait = TRUE, verbose = TRUE))
+  x <- suppressWarnings(wdpa_fetch("LIE", wait = TRUE, verbose = TRUE))
   expect_is(x, "sf")
   expect_true(all(x$ISO3 == "LIE"))
 })
@@ -22,13 +21,39 @@ test_that("ISO3", {
 test_that("global", {
   skip_on_cran()
   skip_if_not(curl::has_internet())
-  skip_if_not_installed("pingr")
+  skip_if_local_and_slow_internet()
   skip_on_github_workflow("Windows")
   skip_on_github_workflow("Mac OSX")
-  skip_if(
-    (mean(pingr::ping("www.google.com", count = 10)) > 10) &&
-    !identical(Sys.getenv("CI"), "true"))
   x <- suppressWarnings(wdpa_fetch(
-    "global", force = TRUE, wait = TRUE, n = 5, verbose = TRUE))
+    "global", wait = TRUE, n = 5, verbose = TRUE))
   expect_is(x, "sf")
+})
+
+test_that("polygon and point data", {
+  skip_on_cran()
+  skip_if_not(curl::has_internet())
+  skip_if_local_and_slow_internet()
+  skip_on_github_workflow("Windows")
+  skip_on_github_workflow("Mac OSX")
+  x <- suppressWarnings(wdpa_fetch("USA", wait = TRUE))
+  expect_is(x, "sf")
+  expect_true(any(vapply(
+    sf::st_geometry(x), inherits, logical(1), c("POLYGON", "MULTIPOLYGON")
+  )))
+  expect_true(any(vapply(
+    sf::st_geometry(x), inherits, logical(1), c("POINT", "MULTIPOINT")
+  )))
+})
+
+test_that("cache", {
+  skip_on_cran()
+  skip_if_not(curl::has_internet())
+  skip_on_github_workflow("Windows")
+  # fetch data
+  x <- wdpa_fetch("MHL", wait = TRUE, force = TRUE)
+  Sys.sleep(2)
+  y <- wdpa_fetch("MHL", wait = TRUE, force = FALSE)
+  # run tests
+  expect_is(x, "sf")
+  expect_equal(x, y)
 })